Move WidgetRef test to dedicated file
This is in preparation for the crate split, after which TestHarness won't be available from widget_ref.rs anymore.
